chore: rename systemd to daemon and move the linux.go
