+++ client/cmd/daemon/linux_test.go
@@ -1,4 +1,4 @@
-package systemd
+package daemon
 
 import (
 	"os"
@@ -29,26 +29,6 @@ func TestDerivePaths_Defaults(t *testing.T) {
 	up, _ := userPaths(Options{ServiceName: "hoopd"})
 	if !strings.HasSuffix(up, "/.config/systemd/user/hoopd.service") {
 		t.Fatalf("unit path=%q, want ~/.config/systemd/user/hoopd.service", up)
-	}
-}
-
-func withTempHome(t *testing.T) (restore func(), home string) {
-	t.Helper()
-	oldHome := os.Getenv("HOME")
-	td := t.TempDir()
-	if err := os.Setenv("HOME", td); err != nil {
-		t.Fatalf("failed to set HOME to %q: %v", td, err)
-	}
-
-	return func() {
-		_ = os.Setenv("HOME", oldHome)
-	}, td
-}
-
-func requireContains(t *testing.T, s, substr string) {
-	t.Helper()
-	if !strings.Contains(s, substr) {
-		t.Fatalf("expected to contain %q, got:\n%s", substr, s)
 	}
 }
